requirements.txt
scikit-learn==0.21.2
MedPy==0.4.0
scikit-image==0.15.0
joblib==0.13.2
Cython==0.29.7
tqdm==4.29.0
zenlog==1.1
pandas==0.24.2
seaborn==0.9.0
statsmodels==0.9.0
scipy==1.2.1
Flask==1.0.2
numpy==1.16.3
matplotlib==3.0.3
pyOpenSSL==19.0.0
psutil==5.6.2
typing==3.6.6
